backend work completed

+++ Guest_service/src/main/java/com/abhisheksah/repository/Repository.java
@@ -5,7 +5,7 @@ import org.springframework.data.mongodb.repository.Query;
 
 import com.abhisheksah.model.Guest;
 
-public interface Repository extends MongoRepository<Guest, Integer>{
+public interface Repository extends MongoRepository<Guest, Long>{
 
 	/*
 	@Query
